move code from selector to cull-mesh.js
